Cleanup all v1alpha1  in EPP
