+++ components/event-publisher-proxy/pkg/options/options.go
@@ -3,8 +3,6 @@ package options
 import (
 	"flag"
 	"fmt"
-
-	"github.com/kelseyhightower/envconfig"
 )
 
 const (
@@ -14,20 +12,11 @@ const (
 	// All the available arguments.
 	argMaxRequestSize = "max-request-size"
 	argMetricsAddress = "metrics-addr"
-
-	// All the available environment variables.
-	envEnableNewCRDVersion = "ENABLE_NEW_CRD_VERSION"
 )
 
 type Options struct {
 	MaxRequestSize int64
 	MetricsAddress string
-	Env
-}
-
-// Env represents the controller environment variables.
-type Env struct {
-	EnableNewCRDVersion bool `envconfig:"ENABLE_NEW_CRD_VERSION" default:"false"`
 }
 
 func New() *Options {
@@ -39,16 +28,12 @@ func (o *Options) Parse() error {
 	flag.StringVar(&o.MetricsAddress, argMetricsAddress, metricEndpointPort, "The address the metric endpoint binds to.")
 	flag.Parse()
 
-	if err := envconfig.Process("", &o.Env); err != nil {
-		return err
-	}
 	return nil
 }
 
 func (o Options) String() string {
-	return fmt.Sprintf("--%s=%v --%s=%v %s=%v",
+	return fmt.Sprintf("--%s=%v --%s=%v",
 		argMaxRequestSize, o.MaxRequestSize,
 		argMetricsAddress, o.MetricsAddress,
-		envEnableNewCRDVersion, o.EnableNewCRDVersion,
 	)
 }
