+++ components/eventing-controller/pkg/deployment/publisher_deployment.go
@@ -334,10 +334,6 @@ func getBEBEnvVars(publisherConfig env.PublisherConfig) []v1.EnvVar {
 			Name:  "BEB_NAMESPACE",
 			Value: fmt.Sprintf("%s$(BEB_NAMESPACE_VALUE)", bebNamespacePrefix),
 		},
-		{
-			Name:  "ENABLE_NEW_CRD_VERSION",
-			Value: strconv.FormatBool(true),
-		},
 	}
 }
 
@@ -361,7 +357,6 @@ func getNATSEnvVars(natsConfig env.NATSConfig, publisherConfig env.PublisherConf
 		},
 		// JetStream-specific config
 		{Name: "JS_STREAM_NAME", Value: natsConfig.JSStreamName},
-		{Name: "ENABLE_NEW_CRD_VERSION", Value: strconv.FormatBool(true)},
 	}
 }
