+++ resources/eventing/charts/controller/templates/deployment.yaml
@@ -12,10 +12,8 @@ spec:
   template:
     metadata:
       labels: {{- include "controller.selectorLabels" . | nindent 8 }}
-      {{- if .Values.enableNewCRDVersion }}
       annotations:
         traffic.sidecar.istio.io/excludeInboundPorts: {{ .Values.webhook.targetPort | quote }}
-      {{- end }}
     spec:
       serviceAccountName: {{ include "controller.fullname" . }}
       terminationGracePeriodSeconds: 10
@@ -90,8 +88,6 @@ spec:
             value: {{ .Values.webhook.mutating.name | quote }}
           - name: VALIDATING_WEBHOOK_NAME
             value: {{ .Values.webhook.validating.name | quote }}
-          - name: ENABLE_NEW_CRD_VERSION
-            value: {{ .Values.enableNewCRDVersion | quote }}
           resources:
             requests:
               cpu: {{ .Values.resources.requests.cpu }}
@@ -126,24 +122,18 @@ spec:
             - containerPort: {{ .Values.metrics.config.port }}
               name: {{ .Values.global.ports.namePrefix }}{{ .Values.metrics.config.portName }}
               protocol: TCP
-            {{- if .Values.enableNewCRDVersion }}
             - containerPort: {{ .Values.webhook.targetPort }}
               name: webhook-server
               protocol: TCP
-            {{- end }}
-          {{- if .Values.enableNewCRDVersion }}
           volumeMounts:
             - mountPath: /tmp/k8s-webhook-server/serving-certs
               name: cert
               readOnly: true
-          {{- end }}
-      {{- if .Values.enableNewCRDVersion }}
       volumes:
         - name: cert
           secret:
             defaultMode: 420
             secretName: {{ .Values.webhook.secretName }}
-      {{- end }}
     {{- if .Values.global.priorityClassName }}
       priorityClassName: {{ .Values.global.priorityClassName }}
     {{- end }}
