+++ resources/eventing/charts/controller/values.yaml
@@ -102,5 +102,3 @@ jetstream:
   consumerDeliverPolicy: new
   maxMessages: -1 # no limit
   maxBytes: -1
-
-enableNewCRDVersion: true
